# backend/docs/swagger.yaml
definitions:
...
  types.CreateProductCategoryDTO:
    properties:
      description:
        type: string
      name:
        minLength: 2
        type: string
    required:
    - name
    type: object
  types.EmployeeLoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  types.ProductCategoryDTO:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  types.UpdateProductCategoryDTO:
    properties:
      description:
        type: string
      name:
        minLength: 2
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/auth/employees/employee/login:
    post:
      consumes:
      - application/json
      description: Authenticates an employee and returns a session token
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.EmployeeLoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful, session token returned
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid credentials or bad request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Employee account inactive
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Employee login
      tags:
      - employee-auth
...
  /api/v1/product-categories:
    get:
      consumes:
      - application/json
      description: Returns paginated list of product categories
      parameters:
      - description: Search term
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get all product categories
      tags:
      - product-categories
    post:
      consumes:
      - application/json
      description: Creates a new product category
      parameters:
      - description: New category data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.CreateProductCategoryDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create new product category
      tags:
      - product-categories
  /api/v1/product-categories/{id}:
    delete:
      description: Deletes a product category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete product category
      tags:
      - product-categories
    get:
      description: Retrieves product category details by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ProductCategoryDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get product category by ID
      tags:
      - product-categories
    put:
      consumes:
      - application/json
      description: Updates an existing product category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated category data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.UpdateProductCategoryDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a product category
      tags:
      - product-categories
swagger: "2.0"